+++ common/src/main/java/com/github/yajatkaul/mega_showdown/MegaShowdownClient.java
@@ -3,6 +3,7 @@ package com.github.yajatkaul.mega_showdown;
 import com.github.yajatkaul.mega_showdown.block.MegaShowdownBlockEntities;
 import com.github.yajatkaul.mega_showdown.block.block_entity.renderer.PedestalBlockEntityRenderer;
 import com.github.yajatkaul.mega_showdown.client.CobbleClientEvents;
+import com.github.yajatkaul.mega_showdown.networking.client.MegaShowdownNetworkHandlerClient;
 import com.github.yajatkaul.mega_showdown.render.ItemRenderingLoader;
 import com.github.yajatkaul.mega_showdown.render.accessories.AccessoriesRegisterRenderer;
 import dev.architectury.registry.ReloadListenerRegistry;
@@ -12,6 +13,7 @@ import net.minecraft.server.packs.PackType;
 public class MegaShowdownClient {
     public static void init() {
         AccessoriesRegisterRenderer.register();
+        MegaShowdownNetworkHandlerClient.register();
 
         CobbleClientEvents.register();
         ReloadListenerRegistry.register(PackType.CLIENT_RESOURCES, new ItemRenderingLoader());