+++ common/src/main/java/com/github/yajatkaul/mega_showdown/client/WheelDataClient.java
@@ -1,0 +1,9 @@
+package com.github.yajatkaul.mega_showdown.client;
+
+public class WheelDataClient {
+    public static boolean shouldMega = false;
+    public static boolean shouldUltra = false;
+
+    public static boolean canMega = false;
+    public static boolean canUltra = false;
+}