+++ fabric/src/main/java/com/github/yajatkaul/mega_showdown/fabric/MegaShowdownFabric.java
@@ -4,7 +4,6 @@ import com.cobblemon.mod.common.Cobblemon;
 import com.cobblemon.mod.common.battles.runner.graal.GraalShowdownService;
 import com.cobblemon.mod.relocations.graalvm.polyglot.Value;
 import com.github.yajatkaul.mega_showdown.MegaShowdown;
-import com.github.yajatkaul.mega_showdown.cobblemon.features.GlobalFeatureManager;
 import com.github.yajatkaul.mega_showdown.command.MegaShowdownCommands;
 import com.github.yajatkaul.mega_showdown.datapack.MegaShowdownDatapackRegister;
 import com.github.yajatkaul.mega_showdown.fabric.datapack.DatapackRegistry;