+++ neoforge/src/main/java/com/github/yajatkaul/mega_showdown/neoforge/MegaShowdownNeoForgeClient.java
@@ -2,8 +2,6 @@ package com.github.yajatkaul.mega_showdown.neoforge;
 
 import com.github.yajatkaul.mega_showdown.MegaShowdown;
 import com.github.yajatkaul.mega_showdown.MegaShowdownClient;
-import com.github.yajatkaul.mega_showdown.block.MegaShowdownBlockEntities;
-import com.github.yajatkaul.mega_showdown.block.block_entity.renderer.PedestalBlockEntityRenderer;
 import com.github.yajatkaul.mega_showdown.screen.MegaShowdownMenuTypes;
 import com.github.yajatkaul.mega_showdown.screen.custom.ZygardeCubeScreen;
 import net.minecraft.network.chat.Component;
@@ -14,7 +12,6 @@ import net.minecraft.server.packs.repository.PackSource;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
-import net.neoforged.neoforge.client.event.EntityRenderersEvent;
 import net.neoforged.neoforge.client.event.RegisterMenuScreensEvent;
 import net.neoforged.neoforge.event.AddPackFindersEvent;