+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # 1.4+1.7-release
+- Updated for 1.7.1
 - Now sweet max soup and max soup is a cooking pot recipe
 - Mega's devolve on PC placed
 - Delays added to Kyurem Particle Effects (Cries)
 - Now omni ring is fire-resistant
+- Fixed kubfu evolution using any item
+- Improved interaction wheel behaviour
+- Fixed some lang
 
 Fixed Form Headers for:
 - Kyogre